Bug fix for use of describe_stacks
The describe_stacks method in boto will only return 120 stacks. To get
all available stacks the results must be paginated.

Currently if you have over 120 stacks you can create a new stack with
cumulus, but you then cannot update or delete it as the stack will not
be in the results returned by describe_stacks when it is used
unpaginated. This commit fixes this issue.